requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
Flask==3.0.2
h11==0.14.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
moviepy==1.0.3
mutagen==1.47.0
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
pillow==10.2.0
playsound==1.3.0
proglog==0.1.10
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
